feat(ChatbotAttachMenu): Add search
The AttachMenu code is buggy, so I tried a non-ref search here. It took all day but I think this works and I haven't seen the same sorts of bugs. We should go back and suggest this for PF6 as well.
