+++ packages/module/patternfly-docs/content/extensions/virtual-assistant/examples/AttachMenu/AttachMenu.tsx
@@ -5,132 +5,94 @@ import { BellIcon, CodeIcon, ClipboardIcon, CalendarAltIcon, UploadIcon } from '
 import PaperclipIcon from './PaperclipIcon';
 import AttachmentIcon from './AttachmentIcon';
 
+const initialMenuItems = [
+  <DropdownList key="list-1">
+    <DropdownItem value="auth-operator Pod" id="0" icon={<img src={AttachmentIcon} alt="Pod icon" />}>
+      <div className="pf-chatbot__menu-operator">
+        auth-operator
+        <div className="pf-v6-c-menu__item-description">Pod</div>
+      </div>
+    </DropdownItem>
+  </DropdownList>,
+  <DropdownGroup key="group2">
+    <DropdownList>
+      <DropdownItem value="Alerts" id="1" icon={<BellIcon />}>
+        Alerts
+      </DropdownItem>
+      <DropdownItem value="Events" id="2" icon={<CalendarAltIcon />}>
+        Events
+      </DropdownItem>
+      <DropdownItem value="Logs" id="3" icon={<ClipboardIcon />}>
+        Logs
+      </DropdownItem>
+      <DropdownItem value="YAML - Status" id="4" icon={<CodeIcon />}>
+        YAML - Status
+      </DropdownItem>
+      <DropdownItem value="YAML - All contents" id="5" icon={<CodeIcon />}>
+        YAML - All contents
+      </DropdownItem>
+    </DropdownList>
+  </DropdownGroup>
+];
+
+const uploadMenuItems = [
+  <Divider key="divider" />,
+  <DropdownList key="list-2">
+    <DropdownItem key="upload" value="upload" id="upload" icon={<UploadIcon />}>
+      Upload from computer
+    </DropdownItem>
+  </DropdownList>
+];
+
 export const BasicDemo: React.FunctionComponent = () => {
   const [isOpen, setIsOpen] = React.useState<boolean>(false);
-  const [refFullOptions, setRefFullOptions] = React.useState<Element[]>();
-  const [filteredIds, setFilteredIds] = React.useState<string[]>(['*']);
-  const menuRef = React.useRef<HTMLDivElement>(null);
+  const [userFacingMenuItems, setUserFacingMenuItems] = React.useState<React.ReactNode>([]);
 
   const onToggleClick = () => {
-    setIsOpen((prevIsOpen) => !prevIsOpen);
-    setFilteredIds(['*']);
+    setIsOpen(!isOpen);
+    setUserFacingMenuItems(initialMenuItems.concat(uploadMenuItems));
   };
 
-  React.useEffect(() => {
-    if (isOpen && menuRef.current) {
-      const firstElement = menuRef.current.querySelector(
-        'li > button:not(:disabled), li > a:not(:disabled), input:not(:disabled)'
-      );
-      firstElement && (firstElement as HTMLElement).focus();
-      setRefFullOptions(Array.from(menuRef.current.querySelectorAll('li:not(li[role=separator])>*:first-child')));
-    }
-  }, [isOpen]);
+  const findMatchingElements = (elements: React.ReactNode[], targetValue: string) => {
+    let matchingElements = [] as React.ReactNode[];
 
-  const menuItems = [
-    <DropdownList key="group-1">
-      <DropdownItem value="0" id="0" icon={<img src={AttachmentIcon} alt="Pod icon" />}>
-        <div className="pf-chatbot__menu-operator">
-          auth-operator
-          <div className="pf-v6-c-menu__item-description">Pod</div>
-        </div>
-      </DropdownItem>
-    </DropdownList>,
-    <DropdownGroup key="group2">
-      <DropdownList>
-        <DropdownItem value="1" id="1" icon={<BellIcon />}>
-          Alerts
-        </DropdownItem>
-        <DropdownItem value="2" id="2" icon={<CalendarAltIcon />}>
-          Events
-        </DropdownItem>
-        <DropdownItem value="3" id="3" icon={<ClipboardIcon />}>
-          Logs
-        </DropdownItem>
-        <DropdownItem value="4" id="4" icon={<CodeIcon />}>
-          YAML - Status
-        </DropdownItem>
-        <DropdownItem value="5" id="5" icon={<CodeIcon />}>
-          YAML - All contents
-        </DropdownItem>
-      </DropdownList>
-    </DropdownGroup>
-  ];
-
-  const filterItems = (items: any[]) => {
-    if (filteredIds.length === 1 && filteredIds[0] === '*') {
-      return items;
-    }
-
-    const filterChildren = (children: any) => {
-      if (!Array.isArray(children)) {
-        children = [children];
-      }
-      return children.filter((child) => filteredIds.includes(child.props.value));
-    };
-
-    const filterGroup = (group: any) => {
-      if (group.type === DropdownGroup) {
-        const filteredChildren = filterChildren(group.props.children.props.children);
-        if (filteredChildren.length > 0) {
-          return React.cloneElement(group, {
-            children: React.cloneElement(group.props.children, {
-              children: filteredChildren.length > 1 ? filteredChildren : filteredChildren[0]
-            })
-          });
+    elements.forEach((element) => {
+      if (React.isValidElement(element)) {
+        // Check if the element's value matches the targetValue
+        if (element.props.value && element.props.value.toLowerCase().includes(targetValue.toLowerCase())) {
+          matchingElements.push(element);
         }
-      } else if (group.type === DropdownList) {
-        const filteredChildren = filterChildren(group.props.children);
-        if (filteredChildren.length > 0) {
-          return React.cloneElement(group, {
-            children: filteredChildren.length > 1 ? filteredChildren : filteredChildren[0]
-          });
-        }
-      } else if (filteredIds.includes(group.props.value) || group.type === Divider) {
-        return group;
+
+        // Recursively check the element's children
+        const children = React.Children.toArray(element.props.children);
+        matchingElements = matchingElements.concat(findMatchingElements(children, targetValue));
       }
-    };
+    });
 
-    const filteredCopy = items.reduce((acc, group) => {
-      const filteredGroup = filterGroup(group);
-      if (filteredGroup) {
-        acc.push(filteredGroup);
-      }
-      return acc;
-    }, []);
-
-    // Remove the last divider if present
-    if (filteredCopy.length > 0 && filteredCopy[filteredCopy.length - 1].type === Divider) {
-      filteredCopy.pop();
-    }
-
-    return filteredCopy;
+    return matchingElements;
   };
 
   const onTextChange = (textValue: string) => {
     if (textValue === '') {
-      setFilteredIds(['*']);
+      setUserFacingMenuItems(initialMenuItems.concat(uploadMenuItems));
       return;
     }
 
-    const filteredIds =
-      refFullOptions
-        ?.filter((item) => (item as HTMLElement).innerText.toLowerCase().includes(textValue.toString().toLowerCase()))
-        .map((item) => item.id) || [];
-    setFilteredIds(filteredIds);
+    const newMenuItems = findMatchingElements(initialMenuItems, textValue);
+    // this is necessary because the React nodes we find traversing the recursive search
+    // aren't correctly wrapped in a DropdownList. This leads to problems with the
+    // auth-operator item where it winds up floating in a bad place in the DOM and never
+    // gets removed
+    setUserFacingMenuItems(
+      <>
+        <DropdownList>
+          {newMenuItems.map((item) => item)}
+          {newMenuItems.length === 0 && <DropdownItem key="no-items">No results found</DropdownItem>}
+        </DropdownList>
+        {uploadMenuItems.map((item) => item)}
+      </>
+    );
   };
-
-  const filteredItems = filterItems(menuItems);
-  if (filteredItems.length === 0) {
-    filteredItems.push(<DropdownItem key="no-items">No results found</DropdownItem>);
-  }
-  filteredItems.push(<Divider />);
-  filteredItems.push(
-    <DropdownList>
-      <DropdownItem key="upload" value="upload" id="upload" icon={<UploadIcon />}>
-        Upload from computer
-      </DropdownItem>
-    </DropdownList>
-  );
 
   return (
     <div>
@@ -139,11 +101,10 @@ export const BasicDemo: React.FunctionComponent = () => {
       consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
       Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
       <AttachMenu
-        filteredItems={filteredItems}
+        filteredItems={userFacingMenuItems}
         isOpen={isOpen}
         onOpenChange={(isOpen) => setIsOpen(isOpen)}
         onOpenChangeKeys={['Escape']}
-        menuRef={menuRef}
         // eslint-disable-next-line no-console
         onSelect={(_ev, value) => console.log('selected', value)}
         handleTextInputChange={onTextChange}
