+++ packages/module/patternfly-docs/content/extensions/virtual-assistant/examples/ChatbotAttachmentMenu/ChatbotAttachmentMenu.md
@@ -20,7 +20,8 @@ propComponents:
     'MessageBar',
     'ChatbotFootnote',
     'MessageBox',
-    'Message'
+    'Message',
+    'MessageBarWithAttachMenuProps'
   ]
 ---
