+++ packages/module/src/index.ts
@@ -57,9 +57,6 @@ export * from './MessageBar';
 export { default as MessageBox } from './MessageBox';
 export * from './MessageBox';
 
-export { default as Popover } from './Popover';
-export * from './Popover';
-
 export { default as PreviewAttachment } from './PreviewAttachment';
 export * from './PreviewAttachment';
